refactor: reorganize bridge controller
