+++ app/scripts/controllers/bridge/constants.ts
@@ -1,0 +1,13 @@
+import { BridgeControllerState, BridgeFeatureFlagsKey } from './types';
+
+export const DEFAULT_BRIDGE_CONTROLLER_STATE: BridgeControllerState = {
+  bridgeFeatureFlags: {
+    [BridgeFeatureFlagsKey.EXTENSION_SUPPORT]: false,
+    [BridgeFeatureFlagsKey.NETWORK_SRC_ALLOWLIST]: [],
+    [BridgeFeatureFlagsKey.NETWORK_DEST_ALLOWLIST]: [],
+  },
+  srcTokens: {},
+  srcTopAssets: [],
+  destTokens: {},
+  destTopAssets: [],
+};
