+++ app/scripts/controllers/bridge/types.ts
@@ -1,0 +1,32 @@
+import { Hex } from '@metamask/utils';
+import { SwapsTokenObject } from '../../../../shared/constants/swaps';
+
+// Maps to BridgeController function names
+export enum BridgeUserAction {
+  SELECT_SRC_NETWORK = 'selectSrcNetwork',
+  SELECT_DEST_NETWORK = 'selectDestNetwork',
+}
+
+export enum BridgeBackgroundAction {
+  SET_FEATURE_FLAGS = 'setBridgeFeatureFlags',
+}
+
+export enum BridgeFeatureFlagsKey {
+  EXTENSION_SUPPORT = 'extensionSupport',
+  NETWORK_SRC_ALLOWLIST = 'srcNetworkAllowlist',
+  NETWORK_DEST_ALLOWLIST = 'destNetworkAllowlist',
+}
+
+export type BridgeFeatureFlags = {
+  [BridgeFeatureFlagsKey.EXTENSION_SUPPORT]: boolean;
+  [BridgeFeatureFlagsKey.NETWORK_SRC_ALLOWLIST]: Hex[];
+  [BridgeFeatureFlagsKey.NETWORK_DEST_ALLOWLIST]: Hex[];
+};
+
+export type BridgeControllerState = {
+  bridgeFeatureFlags: BridgeFeatureFlags;
+  srcTokens: Record<string, SwapsTokenObject>;
+  srcTopAssets: { address: string }[];
+  destTokens: Record<string, SwapsTokenObject>;
+  destTopAssets: { address: string }[];
+};
