+++ ui/pages/bridge/bridge.util.ts
@@ -2,7 +2,7 @@ import { Hex, add0x } from '@metamask/utils';
 import {
   BridgeFeatureFlagsKey,
   BridgeFeatureFlags,
-} from '../../../app/scripts/controllers/bridge';
+} from '../../../app/scripts/controllers/bridge/types';
 import {
   BRIDGE_API_BASE_URL,
   BRIDGE_CLIENT_ID,
